Refactor to ensure config is not shared across lambda invocations
